# Conflicts:
#	inference-chain/x/bls/types/message_submit_dealer_part.go
#	inference-chain/x/bls/types/message_submit_group_key_validation_signature.go
#	inference-chain/x/bls/types/message_submit_partial_signature.go
#	inference-chain/x/bls/types/message_submit_verification_vector.go
#	inference-chain/x/bls/types/validate_basic_test.go

+++ local-test-net/launch_add_network_node.sh
@@ -35,6 +35,11 @@ if [ -z "$WIREMOCK_PORT" ]; then
   echo "WIREMOCK_PORT is not set, using $WIREMOCK_PORT"
 fi
 
+if [ -z "$P2P_EXTERNAL_ADDRESS" ]; then
+  echo "P2P_EXTERNAL_ADDRESS is not set"
+  exit 1
+fi
+
 project_name="$KEY_NAME"
 
 docker compose -p "$project_name" down -v
